refactor: IOSDevice inside maestro ios driver
