+++ maestro-client/src/main/java/maestro/drivers/IOSDriver.kt
@@ -20,22 +20,38 @@
 package maestro.drivers
 
 import com.github.michaelbull.result.expect
-import com.github.michaelbull.result.getOrThrow
-import com.github.michaelbull.result.onSuccess
+import device.IOSDevice
 import hierarchy.AXElement
-import ios.IOSDevice
 import ios.IOSDeviceErrors
-import maestro.*
+import maestro.Capability
+import maestro.DeviceInfo
+import maestro.Driver
+import maestro.Filters
+import maestro.KeyCode
+import maestro.MaestroException
+import maestro.MediaExt
+import maestro.NamedSource
+import maestro.Point
+import maestro.ScreenRecording
+import maestro.SwipeDirection
+import maestro.TreeNode
 import maestro.UiElement.Companion.toUiElement
 import maestro.UiElement.Companion.toUiElementOrNull
-import maestro.utils.*
+import maestro.ViewHierarchy
+import maestro.toCommonDeviceInfo
+import maestro.utils.Insight
+import maestro.utils.Insights
+import maestro.utils.MaestroTimer
+import maestro.utils.Metrics
+import maestro.utils.MetricsProvider
+import maestro.utils.NoopInsights
+import maestro.utils.ScreenshotUtils
 import okio.Sink
 import okio.source
 import org.slf4j.LoggerFactory
 import util.XCRunnerCLIUtils
 import java.io.File
 import java.net.SocketTimeoutException
-import java.util.UUID
 import kotlin.collections.set
 
 class IOSDriver(
